+++ options/admin.py
@@ -1,5 +1,9 @@
 from django.contrib import admin
-from options.models import Option, UserOption
+
+from options import get_option_model, get_user_option_model
+
+Option = get_option_model()
+UserOption = get_user_option_model()
 
 
 @admin.register(Option)