+++ options/apps.py
@@ -1,19 +1,21 @@
 import logging
-import six
+
 from django.apps import AppConfig
 from django.db.models.signals import post_migrate
 from django.db.utils import IntegrityError
+from django.utils.translation import ugettext_lazy as _
 
-from options.settings import DEFAULT_OPTIONS
+from options import get_option_model
+from options.settings import DEFAULT_CONFIGURATION
 
 logger = logging.getLogger(__name__)
 
 
 def create_default_options(sender, **kwargs):
     """Creates the defaults configuration options if they don't exists."""
-    from options.models import Option
+    Option = get_option_model()
 
-    for key, data in six.iteritems(DEFAULT_OPTIONS):
+    for key, data in DEFAULT_CONFIGURATION.items():
         if not Option.objects.filter(name=key).exists():
             try:
                 Option.objects.create(name=key, **data)
@@ -23,6 +25,7 @@ def create_default_options(sender, **kwargs):
 
 class ConfigurationsConfig(AppConfig):
     name = "options"
+    verbose_name = _("Options")
 
     def ready(self):
         """Connects signals with their managers."""